Actions and connected classes moved to special package
