+++ src/main/java/com/lilangel/execpions/OurNotImplemetedException.java
@@ -1,4 +1,4 @@
-package com.lilangel.model;
+package com.lilangel.execpions;
 
 public class OurNotImplemetedException extends RuntimeException{
 }
