+++ src/main/java/com/lilangel/model/RobotActions/RobotStandStillAction.java
@@ -1,4 +1,6 @@
-package com.lilangel.model;
+package com.lilangel.model.RobotActions;
+
+import com.lilangel.model.RobotActions.RobotAction;
 
 public class RobotStandStillAction implements RobotAction {
     @Override
